+++ src/components/Organisms/Modal/UnsubscribeChannelModal.tsx
@@ -1,7 +1,10 @@
+// hooks
+import useUnsubscribeChannelModal from './UnSubscribeChannelModal.hook';
+
+// components
 import Button from '@/components/Atoms/Modal/Button';
 import Dimmed from '@/components/Atoms/Modal/Dimmed';
 import { ButtonWrapper, Description, Title, Wrapper } from '@/components/Atoms/Modal/StyledComponents';
-import useUnsubscribeChannelModal from './UnSubscribeChannelModal.hook';
 
 export default function UnSubscribeChannelModal() {
   const logics = useUnsubscribeChannelModal();
