+++ src/components/Organisms/Modal/UnsubscribeChannelModal.type.tsx
@@ -1,0 +1,3 @@
+export interface UnSubscribeChannelModalProps {
+  channelId: string;
+}
