+++ src/pages/Main/Channel/_components/SubscribeButton.type.tsx
@@ -1,0 +1,7 @@
+export interface SubscribeButtonProps {
+  channelId: string | undefined;
+}
+
+export interface UseSubscribeButtonProps {
+  channelId: string | undefined;
+}
